requirements.txt
addict==2.4.0
anyio==3.6.2
azure-core==1.26.4
bcrypt==4.0.1
boto3==1.26.114
botocore==1.29.114
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
contourpy==1.0.7
coverage==7.2.5
cryptography==40.0.2
cycler==0.11.0
dnspython==2.3.0
ecdsa==0.18.0
exceptiongroup==1.1.1
fastapi==0.95.1
filelock==3.11.0
fonttools==4.39.3
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
huggingface-hub==0.13.4
idna==3.4
imageio==2.27.0
importlib-resources==5.12.0
iniconfig==2.0.0
isodate==0.6.1
Jinja2==3.1.2
jmespath==1.0.1
jose==1.0.0
kiwisolver==1.4.4
lazy_loader==0.2
lit==16.0.5
MarkupSafe==2.1.2
matplotlib==3.7.1
mock==5.0.2
motor==3.1.2
mpmath==1.3.0
msrest==0.7.1
networkx==3.1
ninja==1.11.1
numpy==1.24.2
# nvidia-cublas-cu11==11.10.3.66
# nvidia-cuda-cupti-cu11==11.7.101
# nvidia-cuda-nvrtc-cu11==11.7.99
# nvidia-cuda-runtime-cu11==11.7.99
# nvidia-cudnn-cu11==8.5.0.96
# nvidia-cufft-cu11==10.9.0.58
# nvidia-curand-cu11==10.2.10.91
# nvidia-cusolver-cu11==11.4.0.1
# nvidia-cusparse-cu11==11.7.4.91
# nvidia-nccl-cu11==2.14.3
# nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.5.5.64
opencv-python-headless==4.5.4.60
packaging==23.1
passlib==1.7.4
Pillow==9.5.0
pluggy==1.0.0
pyasn1==0.5.0
pyclipper==1.3.0.post4
pycocotools==2.0.6
pycparser==2.21
pydantic==1.10.7
pymongo==4.3.3
pyparsing==3.0.9
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-mock==3.10.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyWavelets==1.4.1
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
scikit-image==0.20.0
scipy==1.9.1
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
supervision==0.6.0
sympy==1.11.1
tifffile==2023.4.12
timm==0.6.13
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tqdm==4.65.0
transformers==4.28.1
# triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
yapf==0.32.0
zipp==3.15.0
